pull past messages implemented
