--- app/src/main/java/com/line/fukuokabclient/Adapter/ChannelsAdapter.kt
@@ -1,36 +1,0 @@
-package com.line.fukuokabclient.Adapter
-
-import android.content.Intent
-import android.support.v7.widget.RecyclerView
-import android.view.LayoutInflater
-import android.view.View
-import android.view.ViewGroup
-import com.line.fukuokabclient.*
-import com.line.fukuokabclient.dto.ChannelDTO
-import kotlinx.android.synthetic.main.fragment_channels.view.*
-
-class ChannelsAdapter(private val channels: List<ChannelDTO>, val id: Long): RecyclerView.Adapter<ChannelsAdapter.ViewHolder>() {
-    override fun onCreateViewHolder(parent: ViewGroup, viewType: Int): ViewHolder {
-        val inflatedView = LayoutInflater.from(parent.context).inflate(R.layout.fragment_channels, parent, false)
-        return ViewHolder(inflatedView, id)
-    }
-
-    override fun getItemCount(): Int = channels.size
-
-    override fun onBindViewHolder(holder: ViewHolder, position: Int) {
-        val channel = channels[position]
-        holder.bind(channel)
-    }
-
-    class ViewHolder(itemView: View, val id: Long): RecyclerView.ViewHolder(itemView) {
-        fun bind(channel: ChannelDTO) {
-            itemView.txt_channel_name.text = channel.name
-//            itemView.btn_enter_channel.setOnClickListener {
-//                var intent = Intent(itemView.context, ChatActivity::class.java)
-//                intent.putExtra("id", id)
-//                intent.putExtra("channelId", channel.id)
-//                itemView.context.startActivity(intent)
-//            }
-        }
-    }
-}
