+++ app/src/main/java/com/line/fukuokabclient/ChatActivity.kt
@@ -5,11 +5,16 @@ import android.os.Bundle
 import android.support.v7.widget.LinearLayoutManager
 import android.util.Log
 import com.google.firebase.auth.FirebaseAuth
+import com.google.gson.GsonBuilder
 import com.line.fukuokabclient.Adapter.ChatAdapter
 import com.line.fukuokabclient.Utility.Prefs
+import com.line.fukuokabclient.client.ChannelClient
 import com.line.fukuokabclient.dto.MessageDTO
 import com.line.fukuokabclient.websocket.WebSocketChatClient
 import kotlinx.android.synthetic.main.activity_chat.*
+import retrofit2.Retrofit
+import retrofit2.adapter.rxjava.RxJavaCallAdapterFactory
+import retrofit2.converter.gson.GsonConverterFactory
 import rx.android.schedulers.AndroidSchedulers
 import rx.schedulers.Schedulers
 
@@ -17,8 +22,16 @@ class ChatActivity : AppCompatActivity() {
     private var client = WebSocketChatClient(this)
     var mAuth: FirebaseAuth? = null
     var email:String = ""
-    var channelId: Long = 123
-    var senderId:Long = 9999
+    var channelId: Long = 0
+    var senderId:Long = 0
+
+    val gson = GsonBuilder().create()
+    val retrofit = Retrofit.Builder()
+            .baseUrl(BuildConfig.BASEURL)
+            .addConverterFactory(GsonConverterFactory.create(gson))
+            .addCallAdapterFactory(RxJavaCallAdapterFactory.create())
+            .build()
+    val channelClient = retrofit.create(ChannelClient::class.java)
 
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
@@ -64,23 +77,36 @@ class ChatActivity : AppCompatActivity() {
 
     fun start() {
         var items = ArrayList<MessageDTO>()
-        val messageAdapter = ChatAdapter(items, senderId)
-        chat_recycler_view.layoutManager = LinearLayoutManager(this)
-        chat_recycler_view.adapter = messageAdapter
 
-
-        client.topic("/topic/chat.$channelId")
+        channelClient.getMessages(channelId)
                 .subscribeOn(Schedulers.io())
                 .observeOn(AndroidSchedulers.mainThread())
                 .subscribe({
-                    Log.d("hogehoge", "${items.size}")
-                    items.add(it)
-                    messageAdapter.notifyDataSetChanged()
+                    items = ArrayList(it)
+                    val messageAdapter = ChatAdapter(items, senderId)
+                    chat_recycler_view.layoutManager = LinearLayoutManager(this)
+                    chat_recycler_view.adapter = messageAdapter
+
+                    client.topic("/topic/chat.$channelId")
+                            .subscribeOn(Schedulers.io())
+                            .observeOn(AndroidSchedulers.mainThread())
+                            .subscribe({
+                                Log.d("hogehoge", "${items!!.size}")
+                                items.add(it)
+                                messageAdapter.notifyDataSetChanged()
+                            }, {
+                                Log.e("hogehoge", "error", it)
+                            }, {
+                                Log.d("hogehoge", "completed")
+                            })
                 }, {
-                    Log.e("hogehoge", "error", it)
-                }, {
-                    Log.d("hogehoge", "completed")
+
                 })
+
+
+
+
+
     }
