+++ app/src/main/java/com/line/fukuokabclient/client/ChannelClient.kt
@@ -1,6 +1,7 @@
 package com.line.fukuokabclient.client
 
 import com.line.fukuokabclient.dto.ChannelDTO
+import com.line.fukuokabclient.dto.MessageDTO
 import retrofit2.http.GET
 import retrofit2.http.Path
 import rx.Observable
@@ -11,4 +12,7 @@ interface ChannelClient {
 
     @GET("chat/personal/{userId}/{friendId}")
     fun getPersonalChannel(@Path("userId")userId: Long, @Path("friendId") friendId: Long): Observable<ChannelDTO>
+
+    @GET("chat/messages/{channelId}")
+    fun getMessages(@Path("channelId") channelId: Long): Observable<List<MessageDTO>>
 }
